Add new configure option --with-mixedrefs for mixed pointer mode
If `--with-mixedrefs` is specified in the confirgure command, the
'default' `OPENJ9_LIBS_SUBDIR` is used for the build. The mxdptrs
CMake file corresponding to the build OS is used as the
`OPENJ9_BUILDSPEC`.

The new mxdpts build specs are defined in OpenJ9.

The compressed and full GC libraries generated with the mixed
build work are copied to the `OPENJ9_LIBS_SUBDIR` alongside the
other existing libraries, depending on the build mode used.

    full or mixed mode: copy j9gc_full29, j9gcchk_full29
    compressed or mixed mode: copy j9gc29, j9gcchk29

Signed-off-by: Sharon Wang <sharon-wang-cpsc@outlook.com>